Move code to es.py
